fix: move `importPaletteFromJsonFile()` from `GlobalData.ts` to `GlobalUtils.ts`
